+++ src/Axle/Constants/AxleConstants.cs
@@ -1,0 +1,11 @@
+﻿// (c) Lykke Corporation 2019 - All rights reserved. No copying, adaptation, decompiling, distribution or any other form of use permitted.
+
+namespace Axle.Constants
+{
+    using System;
+
+    public static class AxleConstants
+    {
+        public static readonly string ServiceId = Guid.NewGuid().ToString();
+    }
+}