+++ src/Axle/Services/IHubConnectionService.cs
@@ -2,11 +2,31 @@
 
 namespace Axle.Services
 {
+    using System.Collections.Generic;
     using System.Threading.Tasks;
-    using Axle.Contracts;
+    using Axle.Dto;
+    using Microsoft.AspNetCore.SignalR;
 
     public interface IHubConnectionService
     {
-        Task TerminateConnections(SessionActivityType reason, params string[] connectionIds);
+        Task OpenConnection(
+            HubCallerContext context,
+            string userName,
+            string accountId,
+            string clientId,
+            string accessToken,
+            bool isSupportUser);
+
+        void CloseConnection(string connectionId);
+
+        bool TryGetSessionId(string connectionId, out int sessionId);
+
+        IEnumerable<int> GetAllConnectedSessions();
+
+        IEnumerable<string> FindBySessionId(int sessionId);
+
+        IEnumerable<string> FindByAccessToken(string accessToken);
+
+        Task TerminateConnections(TerminateConnectionReason reason, params string[] connectionIds);
     }
 }