+++ src/Axle/Services/INotificationService.cs
@@ -2,19 +2,12 @@
 
 namespace Axle.Services
 {
-    using System;
-    using Axle.Contracts;
     using Axle.Dto;
 
     public interface INotificationService
     {
-#pragma warning disable CA1710 // Event name should end in EventHandler
-        event Action<TerminateSessionNotification> OnSessionTerminated;
-
-        event Action<int> OnBehalfChanged;
-
         void PublishSessionTermination(TerminateSessionNotification terminateSessionNotification);
 
-        void PublishOnBehalfChange(int sessionId);
+        void PublishOtherTabsTermination(TerminateOtherTabsNotification terminateOtherTabsNotification);
     }
 }